function_app/requirements.txt
azure-functions==1.18.0
azure-identity==1.15.0
azure-kusto-data==4.3.1
azure-appconfiguration==1.5.0
requests==2.31.0
asyncio-throttle==1.0.2
aiohttp==3.9.1
